+++ app/entities/BaseEntity.js
@@ -2,9 +2,6 @@ const { EntitySchema } = require('typeorm');
 const BaseColumns = require('./BaseColumns');
 
 module.exports = new EntitySchema({
-  name: 'District',
-  tableName: 'district',
-  columns: {
-    ...BaseColumns,
-  },
+  name: 'BaseEntity',
+  columns: BaseColumns,
 });