+++ app/entities/Department.js
@@ -1,29 +1,13 @@
 const { EntitySchema } = require('typeorm');
-const {
-  db: { srid },
-} = require('config');
+const BaseColumns = require('./BaseColumns');
 
 module.exports = new EntitySchema({
   name: 'Department',
   tableName: 'department',
   columns: {
-    id: {
-      type: Number,
-      primary: true,
-      generated: true,
-    },
-    name: {
-      type: String,
-    },
+    ...BaseColumns,
     capital_name: {
       type: String,
-    },
-    geom: {
-      type: 'geometry',
-      spatialFeatureType: 'MultiPolygon',
-      srid,
-      nullable: false,
-      select: false,
     },
   },
 });